backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
huggingface-hub==0.16.4
mysql-connector-python==8.1.0
transformers==4.35.0
torch==2.1.0
python-dotenv==1.0.0 